backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth & Encryption
python-jose[cryptography]==3.3.0
cryptography==46.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20

# LLM Providers
anthropic==0.42.0
openai==1.58.1
httpx==0.28.1

# File Processing
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4

# PPTX Generation
python-pptx==1.0.2

# Image Processing (brand color extraction)
Pillow==11.1.0

# Redis
redis==5.2.1

# PDF & DOCX Export
python-docx==1.1.2
weasyprint==60.2
pydyf==0.10.0
Jinja2==3.1.4

# Utilities
python-dotenv==1.0.1
